LayoutTests/crypto/workers/subtle/hmac-import-key-sign-expected.txt

[Worker] Test signing with HMAC SHA-1 using an imported key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/hmac-import-key-sign.js
PASS [Worker] bytesToHexString(signature) is expectedSignature
PASS successfullyParsed is true

TEST COMPLETE
